refactor: #10 change position
