+++ src/main/java/laughcandidate/yellowribbonbe/ai/api/BusinessValidationClient.java
@@ -1,4 +1,4 @@
-package laughcandidate.yellowribbonbe.business.client;
+package laughcandidate.yellowribbonbe.ai.api;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
