+++ src/main/java/laughcandidate/yellowribbonbe/business/config/FeignConfig.java
@@ -4,6 +4,6 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@EnableFeignClients(basePackages = "laughcandidate.yellowribbonbe.business.client")
+@EnableFeignClients(basePackages = {"laughcandidate.yellowribbonbe.ai.api", "laughcandidate.yellowribbonbe.business"})
 public class FeignConfig {
 }
